pyproject.toml
[project]
name = "tg-vc-streamer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "async-timeout==5.0.1",
    "attrs==25.4.0",
    "av==16.1.0",
    "deprecation==2.1.0",
    "dotmap==1.3.30",
    "ffmpeg-python==0.2.0",
    "frozenlist==1.8.0",
    "future==1.0.0",
    "idna==3.11",
    "multidict==6.7.1",
    "ntgcalls==1.3.4",
    "numpy==2.2.6",
    "opencv-python-headless==4.13.0.92",
    "packaging==26.0",
    "pillow==12.1.1",
    "propcache==0.4.1",
    "py-tgcalls==2.1.1",
    "pyaes==1.6.1",
    "pyrogram==2.0.106",
    "pysocks==1.7.1",
    "python-arq==6.0.7",
    "python-dotenv==1.2.1",
    "tgcalls==3.0.0.dev6",
    "tgcrypto==1.2.5",
    "typing-extensions==4.15.0",
    "yarl==1.22.0",
    "yt-dlp==2026.2.9.233747.dev0",
]
